feat: middle save #2 (EV)

+++ kumamc-api/src/main/kotlin/net/projecttl/kuma/mc/api/model/Elevator.kt
@@ -2,4 +2,4 @@ package net.projecttl.kuma.mc.api.model
 
 import net.minestom.server.coordinate.Vec
 
-data class Elevator(val loc: Vec, val floors: MutableList<Int>)
+data class Elevator(val loc: Vec, val floors: MutableList<Double>)
